fix: 修改该组件嵌套位置
