+++ packages/cli/packages/template/qunar/src/app.less
@@ -1,7 +1,7 @@
 @import './components/calendar/index.less';
-@import './components/MultiBaseComponent/index.less';
-@import './components/MultiComponent/index.less';
-@import './components/MultiFatherComponent/index.less';
+@import './components/Son/index.less';
+@import './components/Super/index.less';
+@import './components/Parent/index.less';
 .page-guide-text{ 
     color: #333;
     text-align: center;
