--- packages/cli/packages/template/qunar/src/components/MultiBaseComponent/index.less
@@ -1,3 +1,0 @@
-.multi-base-component{
-    font-size: 25rpx;
-}
