--- packages/cli/packages/template/qunar/src/components/MultiFatherComponent/index.js
@@ -1,22 +1,0 @@
-import React from '@react';
-import MultiBaseComponent from '../MultiBaseComponent/index';
-class MultiFatherComponent extends React.Component {
-	constructor() {
-		super();
-	}
-
-	componentDidMount() {
-		/* eslint-disable */
-		console.log('MultiFatherComponent did mount!');
-	}
-	render() {
-		return (
-			<div class="multi-father-component">
-				<div class="title">第二层组件</div>
-				<MultiBaseComponent />
-			</div>
-		);
-	}
-}
-
-export default MultiFatherComponent;
