+++ packages/cli/packages/template/qunar/src/components/Parent/index.js
@@ -1,0 +1,22 @@
+import React from '@react';
+import Son from '../Son/index';
+class Parent extends React.Component {
+    constructor() {
+        super();
+    }
+
+    componentDidMount() {
+        /* eslint-disable */
+		console.log('Parent did mount!');
+	}
+	render() {
+		return (
+			<div class="parent">
+				<div class="title">第二层组件</div>
+				<Son />
+			</div>
+		);
+	}
+}
+
+export default Parent;
