+++ packages/cli/packages/template/qunar/src/components/Parent/index.less
@@ -1,7 +1,8 @@
-.multi-father-component {
+.parent {
   width: 200rpx;
   height: 200rpx;
-  border: 1px solid rgb(175, 175, 175);
+  border: 1px solid rgb(97, 95, 95);
+  margin-left: 20rpx;
   .title {
     font-size: 30rpx;
   }
