+++ packages/cli/packages/template/qunar/src/components/Son/index.js
@@ -1,0 +1,17 @@
+import React from '@react';
+
+class Son extends React.Component {
+    constructor() {
+        super();
+    }
+
+    componentDidMount() {
+        /* eslint-disable */
+		console.log('Son did mount!');
+	}
+	render() {
+		return <div class="son">最内层组件</div>;
+	}
+}
+
+export default Son;
