+++ packages/cli/packages/template/qunar/src/components/Son/index.less
@@ -1,0 +1,7 @@
+.son{
+    font-size: 25rpx;
+    width: 120rpx;
+    height: 120rpx;
+    border: 1px solid rgb(97, 95, 95);
+    margin-left: 20rpx;
+}
