+++ packages/cli/packages/template/qunar/src/components/Super/index.less
@@ -1,6 +1,6 @@
-.multi-component {
-    width: 250rpx;
-    height: 250rpx;
+.super {
+    width: 280rpx;
+    height: 280rpx;
     border: 1px solid rgb(97, 95, 95);
     title {
         font-size: 35rpx;
