--- packages/cli/packages/template/qunar/src/pages/demo/moultiComponent/index.js
@@ -1,15 +1,0 @@
-import React from '@react';
-import MultiComponent from '@components/MultiComponent/index';
-class P extends React.Component {
-	constructor() {
-		super();
-	}
-	render() {
-		return (
-			<div class="mouti-compoent-wrapper">
-				<MultiComponent />
-			</div>
-		);
-	}
-}
-export default P;
