+++ packages/cli/packages/template/qunar/src/pages/demo/syntax/multiple/index.js
@@ -1,0 +1,15 @@
+import React from '@react';
+import Super from '@components/Super/index';
+class P extends React.Component {
+    constructor() {
+        super();
+    }
+    render() {
+        return (
+            <div class="mouti-compoent-wrapper">
+                <Super />
+            </div>
+        );
+    }
+}
+export default P;
